requirements.txt
browserid
pytest==2.7.3
pytest-selenium
pytest-variables
requests==2.8.1
